Move timeline server to separate module
