+++ crates/behavior_simulator/src/main.rs
@@ -1,31 +1,19 @@
-use std::{
-    io::stdout,
-    sync::Arc,
-    time::{Duration, Instant},
-};
+use std::io::stdout;
 
 use chrono::Local;
-use color_eyre::{eyre::bail, install, Result};
-use cycler::Database;
-use fern::Dispatch;
-use framework::{multiple_buffer_with_slots, Reader, Writer};
-
+use color_eyre::{install, Result};
+use fern::{Dispatch, InitError};
 use log::LevelFilter;
-use serde::{Deserialize, Serialize};
-use serialize_hierarchy::SerializeHierarchy;
-use tokio::{select, sync::Notify, time::interval};
 use tokio_util::sync::CancellationToken;
-use types::FieldDimensions;
 
 mod cycler;
 mod interfake;
 mod robot;
+mod server;
 mod simulator;
 mod state;
 
-use crate::simulator::Simulator;
-
-fn setup_logger(is_verbose: bool) -> Result<(), fern::InitError> {
+fn setup_logger(is_verbose: bool) -> Result<(), InitError> {
     Dispatch::new()
         .format(|out, message, record| {
             out.finish(format_args!(
@@ -46,163 +34,6 @@ fn setup_logger(is_verbose: bool) -> Result<(), fern::InitError> {
     Ok(())
 }
 
-#[derive(Clone, Serialize, Deserialize, SerializeHierarchy)]
-struct Configuration {
-    frame: usize,
-    field_dimensions: FieldDimensions,
-}
-
-#[derive(Clone, Default, Serialize, Deserialize, SerializeHierarchy)]
-struct MainOutputs {
-    frame_count: usize,
-    databases: Vec<Database>,
-}
-
-#[derive(Clone, Default, Serialize, Deserialize, SerializeHierarchy)]
-struct BehaviorSimulatorDatabase {
-    main_outputs: MainOutputs,
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn timeline_server(
-    keep_running: CancellationToken,
-    parameters_reader: Reader<Configuration>,
-    parameters_changed: Arc<Notify>,
-    outputs_writer: Writer<BehaviorSimulatorDatabase>,
-    outputs_changed: Arc<Notify>,
-    control_writer: Writer<Database>,
-    control_changed: Arc<Notify>,
-    frames: Vec<Vec<Database>>,
-) {
-    // Hack to provide frame count to clients initially.
-    // Can be removed if communication sends data for
-    // subscribed outputs immediately after subscribing
-    let mut interval = interval(Duration::from_secs(1));
-
-    loop {
-        select! {
-            _ = parameters_changed.notified() => { }
-            _ = interval.tick() => { }
-            _ = keep_running.cancelled() => {
-                break
-            }
-        }
-
-        let frame = parameters_reader.next().frame;
-
-        {
-            let mut outputs = outputs_writer.next();
-            outputs.main_outputs.frame_count = frames.len();
-            outputs.main_outputs.databases = frames[frame].clone();
-        }
-        outputs_changed.notify_waiters();
-
-        {
-            let mut control = control_writer.next();
-            *control = frames[frame][2].clone();
-        }
-        control_changed.notify_waiters();
-    }
-}
-
-fn run(keep_running: CancellationToken) -> Result<()> {
-    let communication_server = communication::server::Runtime::<Configuration>::start(
-        Some("[::]:1337"),
-        "tools/behavior-simulator",
-        "behavior_simulator".to_string(),
-        "behavior_simulator".to_string(),
-        2,
-        keep_running.clone(),
-    )?;
-
-    let (outputs_writer, outputs_reader) =
-        multiple_buffer_with_slots([Default::default(), Default::default(), Default::default()]);
-
-    let outputs_changed = Arc::new(Notify::new());
-    let (subscribed_outputs_writer, _subscribed_outputs_reader) =
-        multiple_buffer_with_slots([Default::default(), Default::default(), Default::default()]);
-
-    communication_server.register_cycler_instance(
-        "BehaviorSimulator",
-        outputs_changed.clone(),
-        outputs_reader,
-        subscribed_outputs_writer,
-    );
-
-    let (control_writer, control_reader) =
-        multiple_buffer_with_slots([Default::default(), Default::default(), Default::default()]);
-
-    let control_changed = Arc::new(Notify::new());
-    let (subscribed_control_writer, _subscribed_control_reader) =
-        multiple_buffer_with_slots([Default::default(), Default::default(), Default::default()]);
-    communication_server.register_cycler_instance(
-        "Control",
-        control_changed.clone(),
-        control_reader,
-        subscribed_control_writer,
-    );
-
-    let mut simulator = Simulator::new();
-    simulator.execute_script("test.lua");
-    let mut frames = Vec::new();
-
-    let start = Instant::now();
-    loop {
-        simulator.cycle();
-
-        let state = simulator.state.lock();
-        let robot_frames = state
-            .robots
-            .iter()
-            .map(|robot| robot.database.clone())
-            .collect();
-        frames.push(robot_frames);
-
-        if state.finished {
-            break;
-        }
-    }
-    let duration = Instant::now() - start;
-    println!("Took {:.2} seconds", duration.as_secs_f32());
-
-    let runtime = tokio::runtime::Runtime::new()?;
-    {
-        let parameters_changed = communication_server.get_parameters_changed();
-        let parameters_reader = communication_server.get_parameters_reader();
-        runtime.spawn(async {
-            timeline_server(
-                keep_running,
-                parameters_reader,
-                parameters_changed,
-                outputs_writer,
-                outputs_changed,
-                control_writer,
-                control_changed,
-                frames,
-            )
-            .await
-        });
-    }
-
-    let mut encountered_error = false;
-    match communication_server.join() {
-        Ok(Err(error)) => {
-            encountered_error = true;
-            println!("{error:?}");
-        }
-        Err(error) => {
-            encountered_error = true;
-            println!("{error:?}");
-        }
-        _ => {}
-    }
-
-    if encountered_error {
-        bail!("at least one cycler exited with error");
-    }
-    Ok(())
-}
-
 fn main() -> Result<()> {
     setup_logger(true)?;
     install()?;
@@ -214,5 +45,6 @@ fn main() -> Result<()> {
             keep_running.cancel();
         })?;
     }
-    run(keep_running)
+
+    server::run(keep_running)
 }
